# backend/internal/docs/swagger.yaml
definitions:
...
  handler.PublishCommentRequest:
    properties:
      content:
        type: string
      parent_id:
        type: integer
      root_id:
        type: integer
    required:
    - content
    type: object
...
  response.PageResponse:
    properties:
      has_more:
        type: boolean
      items: {}
      next_cursor:
        type: string
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
info:
  contact: {}
  description: 高性能短视频 Feed 流系统 API 文档
  termsOfService: http://swagger.io/terms/
  title: TideFlow 短视频 Feed 流 API
  version: v1
paths:
...
  /api/v1/feed/following:
    get:
      description: 获取当前用户关注的人的发布视频
      parameters:
      - description: 游标
        in: query
        name: cursor
        type: string
      - default: 20
        description: 每页数量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - OAuth2Password: []
      summary: 关注流
      tags:
      - 时间线
  /api/v1/feed/latest:
    get:
      description: 获取全局最新视频列表，按时间倒序
      parameters:
      - description: 游标
        in: query
        name: cursor
        type: string
      - default: 20
        description: 每页数量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PageResponse'
      summary: 全局最新Feed
      tags:
      - 时间线
  /api/v1/feed/popular:
    get:
      description: 获取热门视频列表，按热度降序
      parameters:
      - default: "0"
        description: 偏移量游标
        in: query
        name: cursor
        type: string
      - default: 20
        description: 每页数量
        in: query
        name: limit
        type: integer
      - default: 1h
        description: 时间窗口
        in: query
        name: window
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PageResponse'
      summary: 热门榜单
      tags:
      - 时间线
  /api/v1/feed/tag:
    get:
      description: 通过标签搜索视频列表
      parameters:
      - description: 标签名
        in: query
        name: tag
        required: true
        type: string
      - description: 游标
        in: query
        name: cursor
        type: string
      - default: 20
        description: 每页数量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: 按标签搜索视频
      tags:
      - 时间线
  /api/v1/likes/mine:
    get:
      description: 获取当前用户点赞过的视频列表
      parameters:
      - description: 游标
        in: query
        name: cursor
        type: string
      - default: 20
        description: 每页数量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - OAuth2Password: []
      summary: 我点赞过的视频
      tags:
      - 互动
...
  /api/v1/videos/{id}/comments:
    get:
      description: 获取视频的评论列表
      parameters:
      - description: 视频ID
        in: path
        name: id
        required: true
        type: integer
      - description: 游标
        in: query
        name: cursor
        type: string
      - default: 20
        description: 每页数量
        in: query
        name: limit
        type: integer
      - description: 根评论ID
        in: query
        name: root_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: 评论列表
      tags:
      - 互动
    post:
      consumes:
      - application/json
      description: 对视频发表评论或回复
      parameters:
      - description: 视频ID
        in: path
        name: id
        required: true
        type: integer
      - description: 评论内容
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.PublishCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - OAuth2Password: []
      summary: 发表评论
      tags:
      - 互动
  /api/v1/videos/{id}/comments/{comment_id}:
    delete:
      description: 删除指定评论，仅评论作者可删除
      parameters:
      - description: 视频ID
        in: path
        name: id
        required: true
        type: integer
      - description: 评论ID
        in: path
        name: comment_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - OAuth2Password: []
      summary: 删除评论
      tags:
      - 互动
  /api/v1/videos/{id}/like:
    delete:
      description: 取消对指定视频的点赞
      parameters:
      - description: 视频ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - OAuth2Password: []
      summary: 取消点赞
      tags:
      - 互动
    get:
      description: 检查当前用户是否点赞了指定视频
      parameters:
      - description: 视频ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - OAuth2Password: []
      summary: 检查是否点赞
      tags:
      - 互动
    post:
      description: 给指定视频点赞，幂等操作
      parameters:
      - description: 视频ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - OAuth2Password: []
      summary: 点赞视频
      tags:
      - 互动
...
securityDefinitions:
  OAuth2Password:
    flow: password
    scopes:
      read: Grant read access
      write: Grant write access
    tokenUrl: /api/v1/auth/oauth2/token
    type: oauth2
swagger: "2.0"